feat: switched to root org id pattern

+++ backend/src/db/migrations/20251018061215_sub-org.ts
@@ -6,8 +6,12 @@ export async function up(knex: Knex): Promise<void> {
   const hasParentOrgId = await knex.schema.hasColumn(TableName.Organization, "parentOrgId");
   if (!hasParentOrgId) {
     await knex.schema.alterTable(TableName.Organization, (t) => {
+      // the one just above the chain
       t.uuid("parentOrgId");
       t.foreign("parentOrgId").references("id").inTable(TableName.Organization).onDelete("CASCADE");
+      // this would root organization containing various informations like billing etc
+      t.uuid("rootOrgId");
+      t.foreign("rootOrgId").references("id").inTable(TableName.Organization).onDelete("CASCADE");
     });
   }
 
@@ -22,9 +26,11 @@ export async function up(knex: Knex): Promise<void> {
 
 export async function down(knex: Knex): Promise<void> {
   const hasParentOrgId = await knex.schema.hasColumn(TableName.Organization, "parentOrgId");
-  if (hasParentOrgId) {
+  const hasRootOrgId = await knex.schema.hasColumn(TableName.Organization, "rootOrgId");
+  if (hasParentOrgId || hasRootOrgId) {
     await knex.schema.alterTable(TableName.Organization, (t) => {
-      t.dropColumn("parentOrgId");
+      if (hasParentOrgId) t.dropColumn("parentOrgId");
+      if (hasRootOrgId) t.dropColumn("rootOrgId");
     });
   }
